+++ attribution/tests/ddd/repositories/test_load_attribution.py
@@ -82,16 +82,16 @@ class TestLoadAttribution(TestCase):
             self.assertTrue(isinstance(attribution, Attribution))
 
         teacher_attribution = attributions[0].teacher
-        self.assertEqual(attributions[0].acronym, self.l_unit_1_identity.code)
-        self.assertEqual(attributions[0].year, self.l_unit_1_identity.year)
+        self.assertEqual(attributions[0].learning_unit_year.code, self.l_unit_1_identity.code)
+        self.assertEqual(attributions[0].learning_unit_year.year, self.l_unit_1_identity.year)
         self.assertEqual(teacher_attribution.last_name, "Marchal")
         self.assertEqual(teacher_attribution.first_name, "Cali")
         self.assertIsNone(teacher_attribution.middle_name)
         self.assertEqual(teacher_attribution.email, "cali@gmail.com")
 
         teacher_attribution = attributions[1].teacher
-        self.assertEqual(attributions[1].acronym, self.l_unit_1_identity.code)
-        self.assertEqual(attributions[1].year, self.l_unit_1_identity.year)
+        self.assertEqual(attributions[1].learning_unit_year.code, self.l_unit_1_identity.code)
+        self.assertEqual(attributions[1].learning_unit_year.year, self.l_unit_1_identity.year)
         self.assertEqual(teacher_attribution.last_name, "Marchal")
         self.assertEqual(teacher_attribution.first_name, "Tilia")
         self.assertIsNone(teacher_attribution.middle_name)
